requirement.txt
# requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
langchain==0.3.0
langchain-community==0.3.0
langchain-huggingface==0.2.0
langchain-openai==0.2.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.34
sentence-transformers==3.1.1
beautifulsoup4==4.12.3
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.9.2
pgvector==0.3.6
transformers==4.51.3 
langchain_postgres==0.0.14
langchain-google-genai==1.0.8
grpcio==1.71.0